requirements.txt
yfinance
python-dotenv
requests
flask
selenium
scipy
ta